+++ src/MetaheuristicsPlatform/Algorithms/Matheuristics/DINS/DinsMatheuristicOptimizerReferences.cs
@@ -1,0 +1,14 @@
+using MetaheuristicsPlatform.Classification;
+
+namespace MetaheuristicsPlatform.Algorithms.Matheuristics.DINS;
+
+public static class DinsMatheuristicOptimizerReferences
+{
+    public static ScientificReference Primary { get; } =
+        new(
+            "Ghosh",
+            2007,
+            "DINS, a MIP Improvement Heuristic",
+            "IPCO 2007, LNCS",
+            "10.1007/978-3-540-72792-7_24");
+}